pyproject.toml
[tool.poetry]
name = "slashtilities"
version = "0.1.0"
description = "A nice and simple, slash command-based, Discord utility bot."
authors = ["Bryan Hu <bryan.hu.2020@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.6.0"
discord-py-slash-command = "^1.1.0"
disputils = "^0.1.3"
rich = "^10.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.812"
pylint = "^2.7.4"
"discord.py-stubs" = "^1.6.3"
black = "^20.8b1"
isort = "^5.8.0"
poethepoet = "^0.10.0"

[tool.poe.tasks]
    [tool.poe.tasks.run]
    shell = "if [ -r .env ]; then eval $(grep -v '^#' .env | xargs) poetry run python -m slashtilities; else echo 'You need to make a .env'; fi"
    help = "Run the bot"
    # How to kill (requires `ret`. https://pypi.org/projects/ret)
    # `ps aux | egrep  "poe run$" | ret "\w+\s*([^\s]+)" s -g 1 | xargs kill`

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
